refactor(cli): isolate manifest convergence entrypoint
